Fixing that Core.Containers "Show on admin menu" functionality shouldn't depend on the Orchard.Lists feature
